requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
tortoise-orm>=0.17.8
aiosqlite>=0.17.0 